Move to homeassistant.backports.enum
